backend/requirements.txt
# Core API
fastapi
uvicorn[standard]
pydantic
python-multipart

# CAD & Data Processing
ezdxf
pandas
numpy
# ifcopenshell # Uncomment when we add IFC support (complex install in docker)

# Database & Storage
supabase
python-dotenv
httpx

# AI
google-generativeai

# AWS/R2 Storage (S3-compatible)
boto3

# Utilities
loguru
